Concurrent download of actions to improve performance
When downloading actions from outside the US the network latency is
significant. For example, from Norway, a single HTTP request to the
github infrastructure takes 500 milliseconds.

This means the "Set up job" step will be much faster with concurrent
downloads. This change reduces the time for this step from 4 seconds to
1 second in an experiment with 5 action repositories downloaded.